annotations will not be modules atm
better have all the definitions of annot in one file
