--- src/app/annotations/conllu/conllu.module.ts
@@ -1,15 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule
-  ],
-  exports: [
-
-  ]
-})
-export class ConlluModule { }
